+++ app/src/main/java/com/sefford/artdrian/common/ui/Placeholder.kt
@@ -1,0 +1,3 @@
+package com.sefford.artdrian.common.ui
+
+interface Placeholder
