+++ app/src/main/java/com/sefford/artdrian/common/ui/Previews.kt
@@ -1,0 +1,9 @@
+package com.sefford.artdrian.common.ui
+
+import android.content.res.Configuration
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.tooling.preview.Preview
+
+@Preview(name = "Light Mode", uiMode = Configuration.UI_MODE_NIGHT_NO, showBackground = true)
+@Preview(name = "Dark Mode", uiMode = Configuration.UI_MODE_NIGHT_YES, showBackground = true)
+annotation class LightDarkPreview
